+++ lib/src/effects/takeLeading.dart
@@ -37,7 +37,7 @@ part of redux_saga;
 ///
 ///  //...
 ///
-///  fetchUser(action) sync* {
+///  fetchUser({dynamic action}) sync* {
 ///    // ...
 ///  }
 ///
@@ -89,12 +89,10 @@ Iterable<Effect> _TakeLeading(Function saga,
 
     yield Call(saga,
         args: args,
-        namedArgs: _functionHasActionArgument(saga)
-            ? <Symbol, dynamic>{
-                ...?namedArgs,
-                #action: action is Result ? action.value : action
-              }
-            : namedArgs,
+        namedArgs: <Symbol, dynamic>{
+          ...?namedArgs,
+          #action: action is Result ? action.value : action
+        },
         Catch: Catch,
         Finally: Finally);
   }