+++ lib/src/functions.dart
@@ -1,22 +1,5 @@
 part of redux_saga;
 
-///Returns true for the following closures
-///() => void
-///() => Null
-bool _isFunctionVoid(Function f) {
-  if (f == null) throw CannotDetermineNullFunctionReturnType();
-  var closure = f.toString();
-  return closure.contains(') => void') ||
-      closure.contains(') => Null') ||
-      closure.contains(') => Future<void>') ||
-      closure.contains(') => Future<Null>');
-}
-
-bool _functionHasActionArgument(Function f) {
-  var closure = f.toString();
-  return closure.contains(' action}') || closure.contains(' action,');
-}
-
 dynamic _callFunctionWithArgument(Function f, List<dynamic> args,
     Map<Symbol, dynamic> namedArgs, dynamic firstArg) {
   return Function.apply(f, <dynamic>[firstArg, ...?args], namedArgs);
@@ -31,18 +14,6 @@ dynamic _callFinallyFunction(Function f) {
   return Function.apply(f, null);
 }
 
-///possible closures :
-///(Exception) => void
-///(Exception, StackTrace) => void
 dynamic _callErrorFunction(Function f, _SagaInternalException error) {
-//  _checkFunction(f);
-  var closure = f.toString();
-  final args = <dynamic>[];
-  if (closure.contains('(dynamic)')) {
-    args.add(error.message);
-  } else if (closure.contains('(dynamic, StackTrace)')) {
-    args.add(error.message);
-    args.add(error.stackTrace);
-  }
-  return Function.apply(f, args, null);
+  return Function.apply(f, <dynamic>[error.message, error.stackTrace], null);
 }