+++ test/future_test.dart
@@ -20,15 +20,17 @@ void main() {
           actual.add(result.value);
 
           var result2 = Result<dynamic>();
-          yield Call(() => Future<int>.error(exceptionToBeCaught), result: result2);
+          yield Call(() => Future<int>.error(exceptionToBeCaught),
+              result: result2);
           actual.add(result2.value);
-        }, Catch: (dynamic e) sync* {
+        }, Catch: (dynamic e, StackTrace s) sync* {
           actual.add(e);
         });
       });
 
       // saga should handle future resolved/rejected values
-      expect(task.toFuture().then((dynamic value) => actual), completion([1, exceptionToBeCaught]));
+      expect(task.toFuture().then((dynamic value) => actual),
+          completion([1, exceptionToBeCaught]));
     });
   });
 }