+++ test/onerror_test.dart
@@ -9,7 +9,8 @@ void main() {
       dynamic error;
       String stack;
 
-      var sagaMiddleware = createMiddleware(options: Options(onError: (dynamic e, String s) {
+      var sagaMiddleware =
+          createMiddleware(options: Options(onError: (dynamic e, String s) {
         error = e;
         stack = s;
       }));
@@ -26,20 +27,24 @@ void main() {
       expect(task.toFuture(), throwsA(exceptionToBeCaught));
 
       //middleware on error must be invoked
-      expect(task.toFuture().catchError((dynamic e) => error), completion(exceptionToBeCaught));
+      expect(task.toFuture().catchError((dynamic e) => error),
+          completion(exceptionToBeCaught));
 
       expect(
           task.toFuture().catchError((dynamic e) =>
               stack != null &&
-              stack.contains('#0      The above error occurred in task child') &&
+              stack
+                  .contains('#0      The above error occurred in task child') &&
               stack.contains('#1       created by main')),
           completion(true));
     });
 
-    test('saga onError is called for uncaught error (thrown Error instance)', () {
+    test('saga onError is called for uncaught error (thrown Error instance)',
+        () {
       dynamic actual;
 
-      var sagaMiddleware = createMiddleware(options: Options(onError: (dynamic e, String s) {
+      var sagaMiddleware =
+          createMiddleware(options: Options(onError: (dynamic e, String s) {
         actual = e;
       }));
       var store = createStore(sagaMiddleware);
@@ -54,17 +59,20 @@ void main() {
       expect(task.toFuture(), throwsA(exceptionToBeCaught));
 
       // saga passes thrown Error instance in onError handler
-      expect(task.toFuture().catchError((dynamic e) => e), completion(exceptionToBeCaught));
+      expect(task.toFuture().catchError((dynamic e) => e),
+          completion(exceptionToBeCaught));
 
       //middleware on error must be invoked
-      expect(task.toFuture().catchError((dynamic e) => actual), completion(exceptionToBeCaught));
+      expect(task.toFuture().catchError((dynamic e) => actual),
+          completion(exceptionToBeCaught));
     });
 
     test('saga onError is not called for caught errors', () {
       dynamic actual;
       dynamic caught;
 
-      var sagaMiddleware = createMiddleware(options: Options(onError: (dynamic e, String s) {
+      var sagaMiddleware =
+          createMiddleware(options: Options(onError: (dynamic e, String s) {
         actual = e;
       }));
       var store = createStore(sagaMiddleware);
@@ -73,8 +81,8 @@ void main() {
       var task = sagaMiddleware.run(() sync* {
         yield Call(() sync* {
           throw exceptionToBeCaught;
-        }, Catch: (dynamic error) {
-          caught = error;
+        }, Catch: (dynamic e, StackTrace s) {
+          caught = e;
         }, name: 'child');
       }, name: 'main');
 
@@ -82,7 +90,8 @@ void main() {
           completion(exceptionToBeCaught));
 
       //saga must not call onError
-      expect(task.toFuture().then<dynamic>((dynamic value) => actual), completion(null));
+      expect(task.toFuture().then<dynamic>((dynamic value) => actual),
+          completion(null));
     });
   });
 }